requirements.txt
yt-dlp>=2024.1.0
faster-whisper>=1.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
notion-client>=2.0.0
streamlit>=1.32.0
